# Improve removeShape and removeAllShapes performance