@VictorLamoine 2015-09-07
@JanHackenberg confirms that it solved the problem for him:
http://www.pcl-users.org/removeShape-performance-tp4018703p4039354.html